chore(connect): move @trezor/rollout into connect
