+++ ci/deploy.yml
@@ -154,11 +154,6 @@ beta connect-common deploy npm:
   script:
     - nix-shell --run "yarn && yarn build:libs && cd ./packages/connect-common && npm publish --tag beta"
 
-beta rollout deploy npm:
-  <<: *npm_deploy_rules
-  script:
-    - nix-shell --run "yarn && yarn build:libs && cd ./packages/rollout && npm publish --tag beta"
-
 beta transport deploy npm:
   <<: *npm_deploy_rules
   script:
@@ -194,13 +189,6 @@ connect-common deploy npm:
   <<: *npm_deploy_rules
   script:
     - nix-shell --run "yarn && yarn build:libs && cd ./packages/connect-common && npm publish"
-
-rollout deploy npm:
-  only:
-    <<: *run_everything_rules
-  <<: *npm_deploy_rules
-  script:
-    - nix-shell --run "yarn && yarn build:libs && cd ./packages/rollout && npm publish"
 
 transport deploy npm:
   only:
