+++ ci/test.yml
@@ -153,28 +153,6 @@ suite desktop manual:
     <<: *run_everything_rules
   when: manual
 
-# @trezor/rollout
-
-# This test should exist only until trezor-connect is in monorepo.
-# It checks whether rollout works with currently released webwallet data
-.rollout-e2e:
-  stage: integration testing
-  script:
-    - yarn install --frozen-lockfile --cache-folder .yarn --prefer-offline
-    - yarn workspace @trezor/utils build:lib
-    - yarn workspace @trezor/rollout test:integration
-
-rollout:
-  extends: .rollout-e2e
-  only:
-    <<: *run_everything_rules
-
-rollout manual:
-  extends: .rollout-e2e
-  except:
-    <<: *run_everything_rules
-  when: manual
-
 # @trezor/transport
 .e2e transport:
   stage: integration testing
