+++ packages/connect/jest.config.js
@@ -3,4 +3,5 @@ module.exports = {
     // TODO: https://github.com/trezor/trezor-suite/issues/5319
     testEnvironment: 'jsdom',
     collectCoverage: true,
+    setupFiles: ['<rootDir>/setupJest.ts'],
 };
