+++ packages/connect/package.json
@@ -22,7 +22,8 @@
     "browser": {
         "./lib/index": "./lib/index-browser",
         "./lib/workers/workers": "./lib/workers/workers-browser",
-        "./lib/utils/assets": "./lib/utils/assets-browser"
+        "./lib/utils/assets": "./lib/utils/assets-browser",
+        "./lib/rollout/utils/fetch": "./lib/rollout/utils/fetch-browser"
     },
     "react-native": {
         "./lib/workers/workers": "./lib/workers/workers-react-native"
@@ -43,16 +44,17 @@
     },
     "dependencies": {
         "@trezor/blockchain-link": "^2.1.3",
-        "@trezor/rollout": "^1.3.2",
         "@trezor/transport": "^1.1.2",
         "@trezor/utils": "^1.0.1",
         "@trezor/utxo-lib": "^1.0.0",
         "bignumber.js": "^9.0.2",
         "bowser": "^2.11.0",
         "cbor-web": "^7.0.6",
+        "cross-fetch": "^3.1.5",
         "events": "^3.3.0",
         "parse-uri": "1.0.7",
         "randombytes": "2.1.0",
+        "runtypes": "^6.5.1",
         "tslib": "^2.3.1"
     },
     "devDependencies": {
