+++ packages/connect/src/rollout/__tests__/getInfo.bootloader.test.ts
