+++ packages/connect/src/rollout/__tests__/getInfo.firmware.test.ts
