+++ packages/connect/src/rollout/__tests__/getInfo.required.test.ts
