+++ packages/connect/src/rollout/__tests__/getInfo.rollout.test.ts
