+++ packages/connect/src/rollout/utils/fetch-browser.ts
