--- packages/rollout/.npmignore
@@ -1,44 +1,0 @@
-# Dev directories
-coverage
-tests
-__tests__
-__mocks__
-jest*
-.rpt2_cache
-.eslintignore
-.eslintrc.js
-.gitignore
-.gilab-ci.yml
-.prettierignore
-.prettierrc
-jest.config.integration.js
-tsconfig.lib.json
-tsconfig.json
-setupJest.ts
-start-integration.ts
-
-# Source files
-src
-integration
-docs
-
-# Editors
-.editorconfig
-.idea
-.vscode
-*.sublime-project
-*.sublime-workspace
-
-# Dependency directory
-node_modules
-yarn.lock
-package-lock.json
-
-# OSX
-.DS_Store
-
-# Logs
-logs
-*.log
-.yarnclean
-libDev
