--- packages/rollout/package.json
@@ -1,40 +1,0 @@
-{
-    "name": "@trezor/rollout",
-    "version": "1.3.2",
-    "author": "Trezor <info@trezor.io>",
-    "homepage": "https://github.com/trezor/trezor-suite/packages/rollout",
-    "keywords": [
-        "Trezor",
-        "trezor-rollout",
-        "trezor firmware update"
-    ],
-    "repository": {
-        "type": "git",
-        "url": "git://github.com/trezor/trezor-suite.git"
-    },
-    "bugs": {
-        "url": "https://github.com/trezor/trezor-suite/issues"
-    },
-    "license": "SEE LICENSE IN LICENSE.md",
-    "description": "Small javascript library to get latest safe firmware version for trezor update.",
-    "main": "lib/index.js",
-    "types": "lib/index.d.ts",
-    "browser": {
-        "./lib/utils/fetch": "./lib/utils/fetch-browser"
-    },
-    "scripts": {
-        "test:unit": "jest",
-        "test:watch": "jest --watch",
-        "test:integration": "ts-node ./start-integration.ts",
-        "lint": "eslint '**/*.{ts,tsx,js}'",
-        "type-check": "tsc --build tsconfig.json",
-        "type-check:watch": "yarn type-check -- --watch",
-        "build:lib": "rimraf lib && tsc --build ./tsconfig.lib.json"
-    },
-    "dependencies": {
-        "@trezor/utils": "^1.0.0",
-        "cross-fetch": "^3.1.5",
-        "runtypes": "^6.5.1"
-    },
-    "devDependencies": {}
-}
