--- packages/rollout/tsconfig.json
@@ -1,10 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "module": "commonjs",
-        "skipLibCheck": false,
-        "outDir": "./libDev"
-    },
-    "include": ["."],
-    "references": [{ "path": "../utils" }]
-}
