--- packages/rollout/tsconfig.lib.json
@@ -1,8 +1,0 @@
-{
-    "extends": "../../tsconfig.lib.json",
-    "compilerOptions": {
-        "target": "esnext",
-        "outDir": "./lib"
-    },
-    "include": ["./src"]
-}
